+++ src/merge_sort.h
@@ -11,7 +11,10 @@ namespace do_sort {
     private:
       vector<int> aux;
 
-      void recursive_msort(vector<int>& v, vector<int>& aux, int low, int high);
+      void recursive_msort(vector<int>& v,
+                           vector<int>& aux,
+                           size_t low,
+                           size_t high);
 
     protected:
       void specific_do_sort();
